Moved breadcrumb link type to the breadcrumb component.
